+++ src/hooks/useOnProgressChange.ts
@@ -2,39 +2,50 @@ import Animated, {
     runOnJS,
     useAnimatedReaction,
 } from 'react-native-reanimated';
-import { DATA_LENGTH } from '../constants';
+import { computedOffsetXValueWithAutoFillData } from '@/utils/computedWithAutoFillData';
 import type { TCarouselProps } from '../types';
 
 export function useOnProgressChange(
     opts: {
         size: number;
+        autoFillData: boolean;
+        loop: boolean;
         offsetX: Animated.SharedValue<number>;
         rawData: TCarouselProps['data'];
     } & Pick<TCarouselProps, 'onProgressChange'>
 ) {
-    const { offsetX, rawData, size, onProgressChange } = opts;
+    const { autoFillData, loop, offsetX, rawData, size, onProgressChange } =
+        opts;
+
+    const rawDataLength = rawData.length;
+
     useAnimatedReaction(
         () => offsetX.value,
         (_value) => {
-            let value = _value;
+            let value = computedOffsetXValueWithAutoFillData({
+                value: _value,
+                rawDataLength,
+                size,
+                autoFillData,
+                loop,
+            });
 
-            if (rawData.length === DATA_LENGTH.SINGLE_ITEM) {
-                value = value % size;
-            }
-
-            if (rawData.length === DATA_LENGTH.DOUBLE_ITEM) {
-                value = value % (size * 2);
+            if (!loop) {
+                value = Math.max(
+                    -((rawDataLength - 1) * size),
+                    Math.min(value, 0)
+                );
             }
 
             let absoluteProgress = Math.abs(value / size);
 
             if (value > 0) {
-                absoluteProgress = rawData.length - absoluteProgress;
+                absoluteProgress = rawDataLength - absoluteProgress;
             }
 
             !!onProgressChange &&
                 runOnJS(onProgressChange)(value, absoluteProgress);
         },
-        [onProgressChange, rawData]
+        [loop, autoFillData, rawDataLength, onProgressChange]
     );
 }